refactor(pages): duplicate code in dump and diff pages
